+++ pkg/cmd/gobarista/commands/billing.go
@@ -17,14 +17,16 @@ import (
 )
 
 var Billing = cli.Command{
-	Name:  "billing",
-	Usage: "Billing periods and bills management",
+	Name:    "billing",
+	Aliases: []string{"b"},
+	Usage:   "Billing periods and bills management",
 	Subcommands: []*cli.Command{
 		&BillingNewPeriod,
 		&BillingAddBill,
 		&BillingClosePeriod,
 		&BillingPeriodSummary,
 		&BillingIssueBills,
+		&BillingConfirmPaymentBills,
 	},
 	Action: func(ctx *cli.Context) (err error) {
 		if err = helpers.SetupDatabase(ctx); err != nil {
@@ -45,16 +47,17 @@ var Billing = cli.Command{
 }
 
 var BillingNewPeriod = cli.Command{
-	Name:      "create-period",
+	Name:      "period-create",
+	Aliases:   []string{"pcr"},
 	Usage:     "Add new billing period",
-	ArgsUsage: "[from(2006-01-02) to(2006-01-02) issued(2006-01-02) total_amount]",
+	ArgsUsage: "[from(2006-01-02) to(2006-01-02) issued(2006-01-02) total_amount amount_per_package]",
 	Action: func(ctx *cli.Context) (err error) {
 		if err = helpers.SetupDatabase(ctx); err != nil {
 			return err
 		}
 
-		if ctx.NArg() != 4 {
-			return fmt.Errorf("error: too few arguments: requires (4), get (%d)", ctx.NArg())
+		if ctx.NArg() != 5 {
+			return fmt.Errorf("error: too few arguments: requires (5), get (%d)", ctx.NArg())
 		}
 
 		DateFrom, err := time.Parse("2006-01-02", ctx.Args().Get(0))
@@ -77,12 +80,23 @@ var BillingNewPeriod = cli.Command{
 			return fmt.Errorf("error: cannot parse float: %v", err)
 		}
 
+		AmountPerPackage, err := strconv.ParseFloat(ctx.Args().Get(4), 32)
+		if err != nil {
+			return fmt.Errorf("error: cannot parse float: %v", err)
+		}
+
+		totalMonths := 1
+		if int(DateTo.Sub(DateFrom).Hours()/24/30) > 0 {
+			totalMonths = int(DateTo.Sub(DateFrom).Hours() / 24 / 30)
+		}
+
 		period := models.Period{
-			DateFrom:    DateFrom,
-			DateTo:      DateTo,
-			DateOfIssue: DateOfIssue,
-			TotalMonths: int(DateTo.Sub(DateFrom).Hours() / 24 / 30),
-			TotalAmount: float32(TotalAmount),
+			DateFrom:         DateFrom,
+			DateTo:           DateTo,
+			DateOfIssue:      DateOfIssue,
+			TotalMonths:      totalMonths,
+			TotalAmount:      float32(TotalAmount),
+			AmountPerPackage: float32(AmountPerPackage),
 		}
 
 		id, err := database.InsertPeriod(period)
@@ -96,7 +110,8 @@ var BillingNewPeriod = cli.Command{
 }
 
 var BillingClosePeriod = cli.Command{
-	Name:      "close-period",
+	Name:      "period-close",
+	Aliases:   []string{"pcl"},
 	Usage:     "Close billing period and calculate remaining values such as total quantity or unit price",
 	ArgsUsage: "[id]",
 	Action: func(ctx *cli.Context) (err error) {
@@ -130,7 +145,7 @@ var BillingClosePeriod = cli.Command{
 			totalQuantity += bill.Quantity
 		}
 
-		unitPrice := period.TotalAmount / float32(totalQuantity)
+		unitPrice := (period.TotalAmount - period.Cash) / float32(totalQuantity)
 
 		err = database.UpdatePeriodOnClose(uint(pid), totalQuantity, unitPrice)
 		if err != nil {
@@ -154,6 +169,7 @@ var BillingClosePeriod = cli.Command{
 
 var BillingPeriodSummary = cli.Command{
 	Name:      "period-summary",
+	Aliases:   []string{"psu"},
 	Usage:     "Get numbers and statistics for given billing period",
 	ArgsUsage: "[id]",
 	Action: func(ctx *cli.Context) (err error) {
@@ -182,14 +198,16 @@ var BillingPeriodSummary = cli.Command{
 			return fmt.Errorf("error: cannot get bills for given period: %v", err)
 		}
 
-		fmt.Printf("ID:     		%d\n", period.ID)
-		fmt.Printf("Period: 		%s - %s\n", period.DateFrom.Format("2006-01-02"), period.DateTo.Format("2006-01-02"))
-		fmt.Printf("Unit price: 	%.2f\n", period.UnitPrice)
-		fmt.Printf("Total quantity: %d\n", period.TotalQuantity)
-		fmt.Printf("Total amount: 	%.2f\n", period.TotalAmount)
-		fmt.Printf("Total months:   %d\n", period.TotalMonths)
-		fmt.Printf("Closed: 		%t\n", period.Closed)
-		fmt.Printf("Total bills: 	%d\n", len(bills))
+		fmt.Printf("ID:                %d\n", period.ID)
+		fmt.Printf("Period:            %s - %s\n", period.DateFrom.Format("2006-01-02"), period.DateTo.Format("2006-01-02"))
+		fmt.Printf("Unit price:        %.2f\n", period.UnitPrice)
+		fmt.Printf("Total quantity:    %d\n", period.TotalQuantity)
+		fmt.Printf("Total amount:      %.2f\n", period.TotalAmount)
+		fmt.Printf("Total amount (wc): %.2f (total - cash)\n", period.TotalAmount)
+		fmt.Printf("Total months:      %d\n", period.TotalMonths)
+		fmt.Printf("Cash:              %.2f\n", period.Cash)
+		fmt.Printf("Closed:            %t\n", period.Closed)
+		fmt.Printf("Total bills:       %d\n", len(bills))
 
 		return nil
 	},
@@ -197,6 +215,7 @@ var BillingPeriodSummary = cli.Command{
 
 var BillingAddBill = cli.Command{
 	Name:      "add-bill",
+	Aliases:   []string{"a"},
 	Usage:     "Add bill for given user and period with specified quantity",
 	ArgsUsage: "[user_id period_id quantity]",
 	Action: func(ctx *cli.Context) (err error) {
@@ -251,6 +270,7 @@ var BillingAddBill = cli.Command{
 		if err != nil {
 			return fmt.Errorf("error: cannot add new bill to billing period: %v", err.Error())
 		}
+
 		logger.Info(fmt.Sprintf("new bill successfully added to billing period for user_id=%d, user_name=%s: new bill id: %d", uid, user.Firstname+" "+user.Lastname, id))
 
 		return nil
@@ -259,6 +279,7 @@ var BillingAddBill = cli.Command{
 
 var BillingIssueBills = cli.Command{
 	Name:      "issue-bills",
+	Aliases:   []string{"i"},
 	Usage:     "Issue all bills for giver billing period",
 	ArgsUsage: "[id]",
 	Action: func(ctx *cli.Context) (err error) {
@@ -307,11 +328,76 @@ var BillingIssueBills = cli.Command{
 				logger.Error(fmt.Sprintf("error: billing e-mail has not been sent for bill_id=%d user_id=%d user_name='%s' user_email:'%s'", bill.ID, user.ID, user.Firstname+" "+user.Lastname, user.Email))
 			} else {
 				logger.Info(fmt.Sprintf("billing e-mail has been sent for bill_id=%d user_id=%d user_name='%s' user_email:'%s'", bill.ID, user.ID, user.Firstname+" "+user.Lastname, user.Email))
+
+				err = database.UpdateBillOnIssued(bill.ID)
+				if err != nil {
+					return fmt.Errorf("error: cannot update bill: bill_id=%d: %v", bill.ID, err)
+				}
+			}
+		}
+
+		return nil
+	},
+}
+
+var BillingConfirmPaymentBills = cli.Command{
+	Name:      "confirm-payment",
+	Aliases:   []string{"c"},
+	Usage:     "Send payment confirmation for all paid bills for given period (e-mail will not be sent for already notified users)",
+	ArgsUsage: "[id]",
+	Action: func(ctx *cli.Context) (err error) {
+		if err = helpers.SetupDatabase(ctx); err != nil {
+			return err
+		}
+
+		if err = helpers.SetupMail(ctx); err != nil {
+			return err
+		}
+
+		if ctx.NArg() != 1 {
+			return fmt.Errorf("error: too few arguments: requires (1), get (%d)", ctx.NArg())
+		}
+
+		pid, err := strconv.ParseUint(ctx.Args().Get(0), 10, 64)
+		if err != nil {
+			return fmt.Errorf("error: cannot parse uint: %v", err)
+		}
+
+		period, err := database.SelectPeriodByID(uint(pid))
+		if errors.Is(err, gorm.ErrRecordNotFound) {
+			return fmt.Errorf("error: billing period not found")
+		} else if err != nil {
+			return fmt.Errorf("error: cannot get billing period: %v", err)
+		}
+
+		if !period.Closed {
+			return fmt.Errorf("error: cannot send payment confirmation: period is not closed")
+		}
+
+		bills, err := database.SelectAllBillsForPeriod(uint(pid))
+		if err != nil {
+			return fmt.Errorf("error: cannot get bills for given period: %v", err)
+		}
+
+		for _, bill := range bills {
+			user, err := database.SelectUserByID(bill.UserID)
+			if errors.Is(err, gorm.ErrRecordNotFound) {
+				return fmt.Errorf("error: user for giver bill is not found: user_id=%d", bill.UserID)
+			} else if err != nil {
+				return fmt.Errorf("error: cannot get user: user_id=%d: %v", bill.UserID, err)
 			}
 
-			err = database.UpdateBillOnIssued(bill.ID)
-			if err != nil {
-				return fmt.Errorf("error: cannot update bill: bill_id=%d: %v", bill.ID, err)
+			if !bill.PaymentConfirmation && bill.Paid {
+				if err = mail.SendPaymentConfirmation(user, period, bill); err != nil {
+					logger.Error(fmt.Sprintf("error: billing e-mail has not been sent for bill_id=%d user_id=%d user_name='%s' user_email:'%s'", bill.ID, user.ID, user.Firstname+" "+user.Lastname, user.Email))
+				} else {
+					logger.Info(fmt.Sprintf("payment confirmation has been sent for bill_id=%d user_id=%d user_name='%s' user_email:'%s'", bill.ID, user.ID, user.Firstname+" "+user.Lastname, user.Email))
+
+					err = database.UpdateBillOnPaymentConfirmation(bill.ID)
+					if err != nil {
+						return fmt.Errorf("error: cannot update bill: bill_id=%d: %v", bill.ID, err)
+					}
+				}
 			}
 		}