+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# v1.2.0
+- Added CI CodeQL and Staticcheck
+- Added `AmountPerPackage` to models.Period
+- Added `PaymentConfirmation` to models.Bill
+- Added `Cash` to models.Period
+- Changed TotalAmount computation
+- Mail package reorganization
+- Added payment confirmation e-mail and command
+- Added command for all customers contacts print
+- Fixed bulk add and switched firstname and lastname
+
 # v1.1.1
 - Version fix